pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tam-workflow"
version = "0.1.0"
description = "Automated TAM Workflow System for Buildkite"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "alembic==1.14.1",
    # HTTP client
    "httpx==0.28.1",
    # Google APIs
    "google-api-python-client==2.159.0",
    "google-auth-oauthlib==1.2.1",
    "google-auth-httplib2==0.2.0",
    # Slack
    "slack-sdk==3.34.0",
    "aiohttp==3.11.16",
    # Notion (via httpx)
    # Linear (via httpx + GraphQL)
    # Claude AI
    "anthropic==0.42.0",
    # Scheduling
    "apscheduler==3.10.4",
    # Encryption
    "cryptography==44.0.0",
    # Retry logic
    "tenacity==9.0.0",
    # Settings / env
    "pydantic-settings==2.7.1",
    "pydantic==2.10.5",
    # File parsing
    "python-docx==1.1.2",
    "PyPDF2==3.0.1",
    # Utilities
    "python-dateutil==2.9.0",
    "structlog==24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx",  # for FastAPI TestClient
    "aiosqlite==0.20.0",  # async SQLite for testing without PostgreSQL
    "ruff==0.9.4",
    "mypy==1.14.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
